+++ CHANGELOG.md
@@ -33,6 +33,7 @@ eval count:           1399 token(s)
 eval duration:        17.698413s
 eval rate:            79.05 tokens/s
 ```
+* Primera implementación del loop de control en `airsim-loop`
 
 
 ## 2026-0625